refactor(environment): keep key value parameters domain agnostic
