+++ resources/js/components/common/KeyValueParameters/KeyValueParameters.vue
@@ -17,12 +17,7 @@ import { AppSwitch } from '@/components/base/switch';
 import { AppTooltipWrapper } from '@/components/base/tooltip';
 import { useKeyValueParameters } from '@/composables/ui/useKeyValueParameters';
 import { type ParameterContract } from '@/interfaces/ui';
-import { useEnvironmentVariablesStore, useValueGeneratorStore } from '@/stores';
-import {
-    createEnvironmentVariablesMap,
-    EnvironmentPlaceholderStatus,
-    getEnvironmentPlaceholderStatus,
-} from '@/utils/request';
+import { useValueGeneratorStore } from '@/stores';
 import { cn } from '@/utils/ui';
 import {
     EyeClosedIcon,
@@ -41,6 +36,7 @@ export interface AppKeyValueParametersProps {
     modelValue?: ParameterContract[];
     freeFormTypes?: boolean;
     class?: HTMLAttributes['class'];
+    getValueInputClass?: (parameter: ParameterContract) => HTMLAttributes['class'];
 }
 
 export interface AppKeyValueParametersEmits {
@@ -55,12 +51,12 @@ const props = withDefaults(defineProps<AppKeyValueParametersProps>(), {
     modelValue: () => [],
     freeFormTypes: false,
     class: undefined,
+    getValueInputClass: undefined,
 });
 
 const emit = defineEmits<AppKeyValueParametersEmits>();
 
 const { openCommand, closeCommand } = useValueGeneratorStore();
-const environmentVariablesStore = useEnvironmentVariablesStore();
 
 /*
  * Refs.
@@ -96,26 +92,6 @@ const {
 
 const shouldShowGeneratorIcon = (index: number, parameter: ParameterContract) => {
     return focusedInputIndex.value === index && parameter.enabled;
-};
-
-const activeVariablesMap = computed(() =>
-    createEnvironmentVariablesMap(
-        environmentVariablesStore.activeCollection?.variables ?? [],
-    ),
-);
-
-const getParameterValueClass = (parameter: ParameterContract) => {
-    const status = getEnvironmentPlaceholderStatus(
-        parameter.value,
-        environmentVariablesStore.activeCollection?.variables ?? [],
-        activeVariablesMap.value,
-    );
-
-    return {
-        'text-destructive': status === EnvironmentPlaceholderStatus.Missing,
-        'text-warning': status === EnvironmentPlaceholderStatus.Empty,
-        'text-primary': status === EnvironmentPlaceholderStatus.Resolved,
-    };
 };
 
 const handleValueInputFocus = (index: number, inputRef: HTMLInputElement) => {
@@ -249,7 +225,7 @@ const handleDeleteParameter = (index: number) => {
                         v-model="parameter.value"
                         placeholder="Value"
                         class="pl-panel h-full flex-1 rounded-none border-0 border-r shadow-none focus:ring-0 focus-visible:ring-0"
-                        :class="getParameterValueClass(parameter)"
+                        :class="props.getValueInputClass?.(parameter)"
                         :disabled="!parameter.enabled"
                         name="kv-value"
                         data-testid="kv-value"
