+++ resources/js/components/domain/Client/Request/RequestBody/RequestBodyFormData.vue
@@ -6,7 +6,13 @@
 import KeyValueParametersBuilder from '@/components/common/KeyValueParameters/KeyValueParameters.vue';
 import { type ParameterContract } from '@/interfaces/ui';
 import { ParameterType } from '@/interfaces/ui/key-value-parameters';
-import { nextTick, ref, watch } from 'vue';
+import { useEnvironmentVariablesStore } from '@/stores';
+import {
+    createEnvironmentVariablesMap,
+    EnvironmentPlaceholderStatus,
+    getEnvironmentPlaceholderStatus,
+} from '@/utils/request';
+import { computed, nextTick, ref, watch } from 'vue';
 
 /*
  * Types & Interfaces.
@@ -24,6 +30,7 @@ export interface AppRequestBodyFormDataEmits {
 
 defineProps<AppRequestBodyFormDataProps>();
 const emit = defineEmits<AppRequestBodyFormDataEmits>();
+const environmentVariablesStore = useEnvironmentVariablesStore();
 
 const model = defineModel<FormData | null>({
     default: () => null,
@@ -99,6 +106,26 @@ const handlePayloadUpdate = (parameters: ParameterContract[]) => {
     emit('update:modelValue', convertParametersArrayToFormData(parameters));
 };
 
+const activeVariables = computed(
+    () => environmentVariablesStore.activeCollection?.variables ?? [],
+);
+const activeVariablesMap = computed(() =>
+    createEnvironmentVariablesMap(activeVariables.value),
+);
+const getValueInputClass = (parameter: ParameterContract) => {
+    const status = getEnvironmentPlaceholderStatus(
+        parameter.value,
+        activeVariables.value,
+        activeVariablesMap.value,
+    );
+
+    return {
+        'text-destructive': status === EnvironmentPlaceholderStatus.Missing,
+        'text-warning': status === EnvironmentPlaceholderStatus.Empty,
+        'text-primary': status === EnvironmentPlaceholderStatus.Resolved,
+    };
+};
+
 /*
  * Watchers.
  */
@@ -133,6 +160,7 @@ watch(
     <KeyValueParametersBuilder
         :model-value="payload"
         :free-form-types="true"
+        :get-value-input-class="getValueInputClass"
         @update:parameters="handlePayloadUpdate"
     />
 </template>
