+++ resources/js/components/domain/Client/Request/RequestHeaders/RequestHeaders.vue
@@ -9,7 +9,12 @@ import type { GeneratorType } from '@/interfaces/http';
 import { type SourceGlobalHeaders } from '@/interfaces/http';
 import { type ParameterContract } from '@/interfaces/ui';
 import { ParameterType } from '@/interfaces/ui/key-value-parameters';
-import { useConfigStore, useRequestStore, useValueGeneratorStore } from '@/stores';
+import { useConfigStore, useEnvironmentVariablesStore, useRequestStore, useValueGeneratorStore } from '@/stores';
+import {
+    createEnvironmentVariablesMap,
+    EnvironmentPlaceholderStatus,
+    getEnvironmentPlaceholderStatus,
+} from '@/utils/request';
 import { generateValueFromType } from '@/utils/value-generator/generateValueFromType';
 import { computed, onBeforeMount, type Ref, ref } from 'vue';
 
@@ -32,6 +37,7 @@ defineProps<AppRequestHeadersProps>();
 const requestStore = useRequestStore();
 const configStore = useConfigStore();
 const valueGeneratorStore = useValueGeneratorStore();
+const environmentVariablesStore = useEnvironmentVariablesStore();
 
 /*
  * State.
@@ -69,6 +75,26 @@ const handleHeadersUpdate = (parameters: ParameterContract[]) => {
     syncHeadersWithPendingRequest(parameters);
 };
 
+const activeVariables = computed(
+    () => environmentVariablesStore.activeCollection?.variables ?? [],
+);
+const activeVariablesMap = computed(() =>
+    createEnvironmentVariablesMap(activeVariables.value),
+);
+const getValueInputClass = (parameter: ParameterContract) => {
+    const status = getEnvironmentPlaceholderStatus(
+        parameter.value,
+        activeVariables.value,
+        activeVariablesMap.value,
+    );
+
+    return {
+        'text-destructive': status === EnvironmentPlaceholderStatus.Missing,
+        'text-warning': status === EnvironmentPlaceholderStatus.Empty,
+        'text-primary': status === EnvironmentPlaceholderStatus.Resolved,
+    };
+};
+
 /*
  * Lifecycle.
  */
@@ -96,6 +122,7 @@ onBeforeMount(() => {
     <KeyValueParametersBuilder
         ref="parametersBuilder"
         :model-value="effectiveHeaders"
+        :get-value-input-class="getValueInputClass"
         @update:parameters="handleHeadersUpdate"
     />
 </template>
